requirements.txt
flask
requests
mindee
openai
gspread
oauth2client
google-api-python-client
google-auth
psycopg2-binary
flask
flask-cors
python-dotenv
boto3
